+++ package.json
@@ -86,7 +86,9 @@
     "lint": "run-p -c -s lint:*",
     "lint:eslint": "echo 'ESLint: TODO'",
     "lint:package": "publint",
-    "compile": "tsc --noEmit",
+    "compile": "run-s -c compile:*",
+    "compile:wxt": "tsc --noEmit",
+    "compile:virtual": "tsc --noEmit",
     "test": "vitest",
     "test:coverage": "vitest run --coverage",
     "prepare": "simple-git-hooks",
