+++ src/virtual/mock-browser.ts
@@ -1,3 +1,3 @@
-import { fakeBrowser as mockBrowser } from '~/testing';
+import { fakeBrowser as mockBrowser } from 'wxt/testing';
 
 export default mockBrowser;
