+++ src/virtual/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "types": ["vite/client", "../types/globals.d.ts"],
+    "paths": {
+      "wxt/*": ["../*"]
+    }
+  }
+}
