+++ src/virtual/unlisted-script-entrypoint.ts
@@ -1,5 +1,5 @@
 import definition from 'virtual:user-unlisted-script';
-import { logger } from '~/client/utils/logger';
+import { logger } from '../sandbox/utils/logger';
 
 (async () => {
   try {
