+++ tsconfig.json
@@ -5,5 +5,5 @@
       "~/*": ["./src/*"]
     }
   },
-  "exclude": ["dist", "demo", "e2e/dist", "templates"]
+  "exclude": ["dist", "demo", "e2e/dist", "templates", "src/virtual"]
 }
